+++ src/Sentry/Internal/BackgroundWorker.cs
@@ -9,13 +9,13 @@ internal class BackgroundWorker : IBackgroundWorker, IDisposable
 {
     private readonly ITransport _transport;
     private readonly SentryOptions _options;
-    // private readonly ConcurrentQueue<Envelope> _queue;
     private readonly ConcurrentQueueLite<Envelope> _queue;
     private readonly int _maxItems;
     private readonly CancellationTokenSource _shutdownSource;
     private readonly SemaphoreSlim _queuedEnvelopeSemaphore;
 
     private volatile bool _disposed;
+    private volatile bool _fatalExceptionEncountered;
     private int _currentItems;
 
     internal event EventHandler? OnFlushObjectReceived;
@@ -28,12 +28,11 @@ internal class BackgroundWorker : IBackgroundWorker, IDisposable
         ITransport transport,
         SentryOptions options,
         CancellationTokenSource? shutdownSource = null,
-        ConcurrentQueue<Envelope>? queue = null)
+        ConcurrentQueueLite<Envelope>? queue = null)
     {
         _transport = transport;
         _options = options;
-        // _queue = queue ?? new ConcurrentQueue<Envelope>();
-        _queue = new ConcurrentQueueLite<Envelope>();
+        _queue = queue ?? new ConcurrentQueueLite<Envelope>();
         _maxItems = options.MaxQueueItems;
         _shutdownSource = shutdownSource ?? new CancellationTokenSource();
         _queuedEnvelopeSemaphore = new SemaphoreSlim(0, _maxItems);
@@ -63,6 +62,12 @@ internal class BackgroundWorker : IBackgroundWorker, IDisposable
         if (_disposed)
         {
             throw new ObjectDisposedException(nameof(BackgroundWorker));
+        }
+
+        if (_fatalExceptionEncountered)
+        {
+            _options.LogDebug("Worker encountered a fatal exception. Discarding envelope.");
+            return false;
         }
 
         var eventId = envelope.TryGetEventId(_options.DiagnosticLogger);
@@ -193,6 +198,8 @@ internal class BackgroundWorker : IBackgroundWorker, IDisposable
         }
         catch (Exception e)
         {
+            _fatalExceptionEncountered = true;
+            _queue.Clear();
             _options.LogFatal(e, "Exception in the background worker.");
             throw;
         }
@@ -372,51 +379,5 @@ internal class BackgroundWorker : IBackgroundWorker, IDisposable
             // Dispose the transport if needed
             (_transport as IDisposable)?.Dispose();
         }
-    }
-}
-
-/// <summary>
-/// This class is purely for testing purposes. It's been hacked together in a short amount of time. Performance is no
-/// doubt terrible and it should in no way be used in production code. It does confirm we have a memory issue with the
-/// <see cref="ConcurrentQueue{T}"/> class however. See https://github.com/getsentry/sentry-dotnet/issues/2516
-/// </summary>
-internal class ConcurrentQueueLite<T>
-{
-    private readonly List<T> _queue = new();
-    private int _listCounter = 0;
-
-    public void Enqueue(T item)
-    {
-        lock (_queue) {
-            _queue.Add (item);
-            _listCounter++;
-        }
-    }
-    public bool TryDequeue([NotNullWhen(true)] out T? item)
-    {
-        item = default;
-        lock (_queue) {
-            if (_listCounter > 0) {
-                item = _queue [0];
-                _queue.RemoveAt (0);
-                _listCounter--;
-            }
-        }
-        return item != null;
-    }
-
-    public int Count => _listCounter;
-
-    public bool IsEmpty => _listCounter == 0;
-
-    public bool TryPeek([NotNullWhen(true)] out T? item)
-    {
-        item = default;
-        lock (_queue) {
-            if (_listCounter > 0) {
-                item = _queue [0];
-            }
-        }
-        return item != null;
     }
 }
